Fixed some stuff with collisions and cleaned up code a bit
