+++ src/main/java/com/epiicthundercat/raft/entity/EntityHook.java
@@ -1,5 +1,7 @@
 package com.epiicthundercat.raft.entity;
 
+import net.minecraft.block.Block;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.effect.EntityLightningBolt;
 import net.minecraft.entity.player.EntityPlayer;
@@ -11,6 +13,8 @@ import net.minecraft.util.math.RayTraceResult;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+
+import static net.minecraft.util.math.RayTraceResult.Type.BLOCK;
 
 public class EntityHook extends EntityThrowable {
 
@@ -39,7 +43,7 @@ public class EntityHook extends EntityThrowable {
 	}
 
 	public EntityHook(World worldIn, DamageSource source, BlockPos pos, EntityPlayer player, FloatBarrel barrel1,
-			PlankEntity plank1, ThatchEntity thatch1, ScrapEntity scrap1) {
+					  PlankEntity plank1, ThatchEntity thatch1, ScrapEntity scrap1) {
 		super(worldIn);
 		sourceDamage = source;
 		posB = pos;
@@ -59,10 +63,15 @@ public class EntityHook extends EntityThrowable {
 	@Override
 	public void onUpdate() {
 		super.onUpdate();
+
 		World world = this.world;
 		int x = this.ticksExisted;
-		if ((this.ticksExisted % 2) == 0) {
+		if (this.ticksExisted < 20 && (this.ticksExisted % 2) == 0) {
 			world.spawnParticle(EnumParticleTypes.DRIP_WATER, this.posX, this.posY, this.posZ, 0.0, 0.0, 0.3);
+		}
+
+		if(!this.world.isRemote && this.shootingEntity == null){
+			this.setDead();
 		}
 
 	}
@@ -80,19 +89,46 @@ public class EntityHook extends EntityThrowable {
 		// getEntityWorld().addWeatherEffect(new
 		// EntityLightningBolt(getEntityWorld(),posX,posY,posZ,false));
 
+
+		if(result.typeOfHit == BLOCK){
+			IBlockState state = this.world.getBlockState(result.getBlockPos());
+			Block block = state.getBlock();
+			if(state.getCollisionBoundingBox(this.world, result.getBlockPos()) == Block.NULL_AABB
+					&& block.canCollideCheck(state, false)){
+				return;
+			}
+		}
+
+		if(this.world.isRemote){
+
+			for (int j = 0; j < 4; ++j) {
+				this.world.spawnParticle(EnumParticleTypes.SWEEP_ATTACK, this.posX, this.posY, this.posZ, 0.0D, 0.0D,
+						0.0D);
+			}
+
+			return;
+		}
 		if (result.entityHit != null && result.entityHit != this.shootingEntity) {
-			result.entityHit.attackEntityFrom(DamageSource.causePlayerDamage((EntityPlayer) this.shootingEntity),
-					10.0F);
-			
-			  if(sourceDamage.getEntity() instanceof EntityPlayer &&
-			  ((EntityPlayer)
-			  sourceDamage.getEntity()).capabilities.isCreativeMode){
-			  barrel.extractItems(world, posB, playerIn);
-			  thatch.extractItems(world, posB, playerIn);
-			  scrap.extractItems(world, posB, playerIn);
-			  plank.extractItems(world, posB, playerIn);
-			  
-			  }
+
+			if (result.entityHit instanceof EntityFishable && this.shootingEntity instanceof EntityPlayer) {
+				EntityFishable entityHit = (EntityFishable) result.entityHit;
+
+				entityHit.extractItems(world, posB, (EntityPlayer) this.shootingEntity);
+				entityHit.setDead();
+			}
+			else{
+				result.entityHit.attackEntityFrom(DamageSource.causePlayerDamage((EntityPlayer) this.shootingEntity), 10.0F);
+			}
+
+
+			/*if(sourceDamage != null && sourceDamage.getEntity() instanceof EntityPlayer &&
+					((EntityPlayer) sourceDamage.getEntity()).capabilities.isCreativeMode){
+				barrel.extractItems(world, posB, playerIn);
+				thatch.extractItems(world, posB, playerIn);
+				scrap.extractItems(world, posB, playerIn);
+				plank.extractItems(world, posB, playerIn);
+
+			}*/
 			 
 			/*if (result.entityHit != null && result.entityHit != this.shootingEntity) {
 				if (result.entityHit instanceof FloatBarrel) {
@@ -106,13 +142,14 @@ public class EntityHook extends EntityThrowable {
 				}*/
 //			}
 
-			for (int j = 0; j < 4; ++j) {
-				this.world.spawnParticle(EnumParticleTypes.SWEEP_ATTACK, this.posX, this.posY, this.posZ, 0.0D, 0.0D,
-						0.0D);
+			if (!this.world.isRemote) {
+				this.setDead();
 			}
 		}
-		if (!this.world.isRemote) {
-			this.setDead();
+		else if(result.entityHit == null) {
+			if (!this.world.isRemote) {
+				this.setDead();
+			}
 		}
 	}
 }
